+++ custom_components/esphome_touch_designer/const.py
@@ -3,6 +3,8 @@ PLATFORMS: list[str] = []
 
 # Service called by ESPHome devices to set a light's RGB (avoids passing a list in homeassistant.action)
 SERVICE_SET_LIGHT_RGB = "set_light_rgb"
+# Service called by ESPHome devices to set a light's white temperature (mireds)
+SERVICE_SET_LIGHT_COLOR_TEMP = "set_light_color_temp"
 
 CONF_WIFI_SSID = "wifi_ssid"                    # suggest "!secret wifi_ssid"
 CONF_WIFI_PASSWORD_SECRET = "wifi_password"     # suggest "!secret wifi_password"